requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.16
gunicorn==23.0.0
packaging==24.1
pdf2image==1.17.0
pillow==10.4.0
pypdf2==3.0.1
pytesseract==0.3.13
sqlparse==0.5.1
typing-extensions==4.12.2
